calculation: more speedy (#37)

+++ src/draw_view_utils.c
@@ -6,13 +6,13 @@
 /*   By: hotph <hotph@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/10/08 12:10:01 by hotph             #+#    #+#             */
-/*   Updated: 2023/10/09 16:14:04 by hotph            ###   ########.fr       */
+/*   Updated: 2023/10/15 15:18:57 by hotph            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "draw.h"
 
-bool	is_map(int x, int y, t_meta *meta)
+bool	is_map(size_t x, size_t y, t_meta *meta)
 {
 	if (x < (meta->tile_size * meta->width_map)
 		&& y < (meta->tile_size * meta->height_map))
@@ -31,11 +31,11 @@ double	get_tex_x(
 		return (modf((wall->y / tile_size), &integer));
 }
 
-int	get_color_from_texture(t_tex tex, double x, double y)
+int	get_color_from_texture(t_tex tex, int x, int y)
 {
 	int		color;
 
-	color = *(int *)(tex.addr + ((int)y * tex.line_length
-				+ (int)x * (tex.bpp / 8)));
+	color = *(int *)(tex.addr + (y * tex.line_length
+				+ x * (tex.bpp / 8)));
 	return (color);
 }
